refactor rule errors
move rule error generation and filtering to separate modules
